djchat/requirements.txt
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
cfgv==3.4.0
channels==4.2.2
click==8.1.8
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
filelock==3.18.0
flake8==7.2.0
h11==0.16.0
httptools==0.6.4
identify==2.6.9
idna==3.10
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pre_commit==4.2.0
psycopg2==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.30.0
watchfiles==1.0.5
websockets==15.0.1
